+++ old-ui/app/unlock.js
@@ -69,7 +69,7 @@ UnlockScreen.prototype.render = function () {
           style: {
             margin: 10,
           },
-        }, 'Unlock'),
+        }, 'Log In'),
       ]),
 
       h('.flex-row.flex-center.flex-grow', [